refactor: move opencode session helpers into providers/opencode
Move ResolveOpencodeSession and the x-opencode-session handling out of
core/network into the opencode provider package, per PR review feedback
(akshaydeo: 'these functions should be in providers/opencode/utils.go').
Pure relocation: package network -> opencode, no logic changes.
