+++ core/providers/opencode/utils_test.go
@@ -1,4 +1,4 @@
-package network
+package opencode
 
 import (
 	"context"
